Profile Update Complete

+++ app/Http/Controllers/Admin/DashboardController.php
@@ -27,7 +27,7 @@ class DashboardController extends Controller
     {
         $this->validate($request, [
             'name' => 'required',
-            'userid' => 'required',
+            'userid' => 'required|unique:users',
             'about' => 'sometimes|max:255',
             'image' => 'sometimes|image|mimes:jpg,png,bmp,jpeg|max:2000'
 
@@ -42,7 +42,7 @@ class DashboardController extends Controller
                 Storage::disk('public')->makeDirectory('user');
             }
             // Delete old Image
-            if (Storage::disk('public')->exists('user/' . $user->image)) {
+            if ($user->image !== 'default.jpg' && Storage::disk('public')->exists('user/' . $user->image)) {
                 Storage::disk('public')->delete('user/' . $user->image);
             }
             // Store
